Rename MainView; update solution file to vs2013
