+++ dnGREP.WPF.sln
@@ -1,6 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
+# Visual Studio 2013
+VisualStudioVersion = 12.0.40629.0
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dnGREP.WPF", "dnGREP.WPF\dnGREP.WPF.csproj", "{CBC7AEC0-59FF-4050-B02D-ADB15359CED4}"
 	ProjectSection(ProjectDependencies) = postProject
 		{7317511F-A32C-44AE-9C87-1E4806E891FB} = {7317511F-A32C-44AE-9C87-1E4806E891FB}
@@ -131,7 +133,6 @@ Global
 		{7317511F-A32C-44AE-9C87-1E4806E891FB} = {177DCA75-DAA1-4097-A3AB-EF944566600D}
 		{6B5159A5-C92C-4203-84DE-273FEB30D07B} = {177DCA75-DAA1-4097-A3AB-EF944566600D}
 		{FA9512EF-562C-4DC1-9CF2-7A8FF896D352} = {177DCA75-DAA1-4097-A3AB-EF944566600D}
-		{DB68B5D3-5E4B-47D0-AA86-83D7C181A1E4} = {177DCA75-DAA1-4097-A3AB-EF944566600D}
 	EndGlobalSection
 	GlobalSection(SharpSetup) = preSolution
 		Version = 1.2
